added 2-logic, split up 1-logic
