+++ bytecodes/00.m
@@ -1,0 +1,4 @@
+push 1
+push 2
+push 3
+pall
