+++ bytecodes/07.m
@@ -1,0 +1,10 @@
+push 1
+push 2
+push 3
+pall
+pop
+pall
+pop
+pall
+pop
+pall
