+++ bytecodes/12.m
@@ -1,0 +1,6 @@
+push 1
+push 2
+push 3
+pall
+add
+pall
