+++ cm
@@ -1,0 +1,1 @@
+gcc -Wall -Werror -Wextra -pedantic -g *.c -o monty
